+++ arch/X86/X86Mapping.c
@@ -6614,13 +6614,13 @@ void X86_get_insn_id(cs_struct *h, cs_insn *insn, unsigned int id)
 
 		if (h->detail) {
 			memcpy(insn->detail->regs_read, insns[i].regs_use, sizeof(insns[i].regs_use));
-			insn->detail->regs_read_count = count_positive(insns[i].regs_use);
+			insn->detail->regs_read_count = (uint8_t)count_positive(insns[i].regs_use);
 
 			memcpy(insn->detail->regs_write, insns[i].regs_mod, sizeof(insns[i].regs_mod));
-			insn->detail->regs_write_count = count_positive(insns[i].regs_mod);
+            insn->detail->regs_write_count = (uint8_t)count_positive(insns[i].regs_mod);
 
 			memcpy(insn->detail->groups, insns[i].groups, sizeof(insns[i].groups));
-			insn->detail->groups_count = count_positive(insns[i].groups);
+            insn->detail->groups_count = (uint8_t)count_positive(insns[i].groups);
 
 			if (insns[i].branch || insns[i].indirect_branch) {
 				// this insn also belongs to JUMP group. add JUMP group
@@ -6645,7 +6645,7 @@ bool X86_insn_check_combine(cs_struct *h, cs_insn *insn)
 	if (insn->id == X86_INS_LOCK || insn->id == X86_INS_REP ||
 		insn->id == X86_INS_REPNE) {
 		// then save this as prev_prefix
-		h->prev_prefix = insn->id;
+		h->prev_prefix = (uint8_t)insn->id;
 		return false;
 	}
 
@@ -6654,8 +6654,7 @@ bool X86_insn_check_combine(cs_struct *h, cs_insn *insn)
 		return true;
 	}
 
-	// neither prefix instruction nor having previous instruction as prefix,
-	// so we cannot combine this with a prefix
+	// cannot combine this with a prefix
 	return false;
 }
